backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.10.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
httpx==0.25.2
pendulum==3.0.0
apscheduler==3.10.4
